update to hmvv v3 beta

+++ shell/runPipelines_nextseq.sh
@@ -3,7 +3,7 @@ export SHELL=/usr/bin/bash
 
 if [ $# -eq 0 ]
 then
-	echo "Usage: pipelineThread.sh"
+	echo "Usage: runPipelines_nextseq.sh"
 	echo "-e environment"
   echo "-u user"
   echo "-p password"
@@ -35,6 +35,10 @@ if test $# -gt 0
 fi
 
 
+#remove previous empty log files
+rm -f $(find /var/pipelines_"$environment"/cron_logs/ -name "nextseq_*.txt" -type f -size -375c)
+
+
 function updateStatus() {
 user=$4
 password=$5
@@ -46,63 +50,111 @@ mysql --user="$user" --password="$password" --database="$database" --execute="$i
 
 date_=`date '+%Y-%m-%d %H:%M:%S'`
 currentdate=$(echo $date_ | sed -e 's/ /_/g' -e 's/:/_/g' -e 's/-/_/g' )
-echo "Running nextseq pipelines cron job at - $currentdate "
+echo " $currentdate    INFO  -  Running nextseq pipelines "
 touch  /var/pipelines_"$environment"/run_files/"nextseq_$currentdate".txt
 sudo chmod 777  /var/pipelines_"$environment"/run_files/"nextseq_$currentdate".txt
 
 
 ################################################################################
-# running nextseq pipeline from sampleAnalysisQueue
+# running nextseq pipeline from pipelineQueue
 ################################################################################
 defaultStatus=0
 instrument='nextseq'
-nextseq_statement="select pipelineQueue.queueID from pipelineQueue join samples on samples.sampleID=pipelineQueue.sampleID join assays on assays.assayID = samples.assayID join instruments on instruments.instrumentID = samples.instrumentID where pipelineQueue.status='$defaultStatus'and instruments.instrumentName ='$instrument' order by pipelineQueue.queueID;"
+nextseq_statement="select pipelineQueue.queueID , samples.runID from pipelineQueue join samples on samples.sampleID=pipelineQueue.sampleID join assays on assays.assayID = samples.assayID join instruments on instruments.instrumentID = samples.instrumentID where pipelineQueue.status='$defaultStatus'and instruments.instrumentName ='$instrument' order by pipelineQueue.queueID;"
 
-echo "   ----- Queued NEXTSEQ jobs in this cronjob are --------"
+echo " $currentdate    INFO  -    Queued NEXTSEQ jobs are"
 
-while  read -r queueID  runID  sampleID instrumentID ;
+ids=()
+
+
+while  read -r queueID runID ;
 do
 
-	echo "queueID is $queueID , user is $user,  database is $environment "
+	echo " $currentdate    INFO  -  Updating queue status for NEXTSEQ queueID - $queueID"
 
 	#update sampleAnalysisQueue table and set status of this queue to 1 i.e started processing
-	updatestatement="UPDATE sampleAnalysisQueue SET status=1 WHERE queueID = $queueID;"
+	updatestatement="UPDATE pipelineQueue SET status=1 WHERE queueID = $queueID;"
 	mysql --user="$user" --password="$password" --database="$environment" --execute="$updatestatement"
 
-
-	## convert bcl2fastq
-	lockdir=/var/pipelines_"$environment"/run_files/nextseq.lock
-	if mkdir "$lockdir" ## mkdir is atomic, file or file variable is not
-	then
-
-		echo "successfully acquired lock: $lockdir"
-		echo "bcl2fastq running for  run - $runID ... "
-		runFolder=$(ls -d /home/$instrumentID/*_"$runID"_*)
-		updateStatus "$queueID" "bcl2fastq" "$environment" "$user"  "$password"
-		bcl2fastq --no-lane-splitting --runfolder-dir $runFolder --output-dir $runFolder/out1  &&  echo " bcl2fastq completed for  run - $runID"
-		sudo rm -rf "$lockdir"
-
-	else
-
-		echo "cannot acquire lock, giving up on $lockdir"
-		echo "nextseq cron job did not complete at - $currentdate"
-		updateStatus "$queueID" "ERROR:Second-bcl2fastq" "$environment" "$user"  "$password"
-		exit 1
-
-	fi
-
 	echo "$queueID"  >> /var/pipelines_"$environment"/run_files/"nextseq_$currentdate".txt
+
+	ids+=("$runID/$queueID")
 
 done < <(mysql --user="$user" --password="$password" --database="$environment" --execute="$nextseq_statement" -N)
 
 
+#uniqueRunIDS=($(printf "%s\n" "${runIDs[@]}" | sort -u))
+
+################################################################################
+# running bcl2fastq for unique runIDs
+################################################################################
+
+for idpair in "${ids[@]}"
+do
+
+	runID=$(echo $idpair | cut -d "/" -f 1)
+
+	queueID=$(echo $idpair | cut -d "/" -f 2)
+
+  fastqStatus=$(mysql --user="$user" --password="$password" --database="$environment" -se "select status from pipelineStatusBcl2Fastq where runID='$runID'")
+
+  echo "checking bcl2fastq for runID - $runID and queueid $queueID and runid fastqStatus '$fastqStatus' --"
+
+	if [ $fastqStatus == "0" ]
+	then
+    echo "running bcl2fastq for runID - $runID and queueid $queueID and runid fastqStatus '$fastqStatus' --"
+		## convert bcl2fastq
+		lockdir=/var/pipelines_"$environment"/run_files/nextseq.lock
+		if mkdir "$lockdir" ## mkdir is atomic, file or file variable is not
+		then
+
+			updateStatus "$queueID" "bcl2fastq_running_now" "$environment" "$user"  "$password"
+
+			echo " $currentdate    INFO  -  NEXTSEQ successfully acquired lock: $lockdir"
+			echo " $currentdate    INFO  -  NEXTSEQ bcl2fastq running for  run - $runID queueID - $queueID ... "
+
+			runFolder=$(ls -d /home/nextseq/*_"$runID"_*)
+
+			sudo /usr/local/bin/bcl2fastq --no-lane-splitting --runfolder-dir $runFolder --output-dir $runFolder/out1 -r 5 -d 5 -p 5 -w 5 &
+
+			wait $!
+
+			updatestatement="UPDATE pipelineStatusBcl2Fastq SET status=1 WHERE runID = $runID;"
+			mysql --user="$user" --password="$password" --database="$environment" --execute="$updatestatement"
+
+			updateStatus "$queueID" "bcl2fastq_completed_now" "$environment" "$user"  "$password"
+
+			sudo rm -rf "$lockdir"
+
+		else
+
+			echo " $currentdate    INFO  -  NEXTSEQ cannot acquire lock for bcl2fastq conversion $runID, giving up on $lockdir"
+			echo " $currentdate    INFO  -  NEXTSEQ nextseq cron job DID NOT complete at - $currentdate"
+
+			#update pipelineQueue table and set status of this queue to 0 i.e try again next time
+			updatestatement="UPDATE pipelineQueue SET pipelineQueue.status=0 WHERE queueID = $queueID;"
+			mysql --user="$user" --password="$password" --database="$environment" --execute="$updatestatement"
+
+			updateStatus "$queueID" "bcl2fastq_wait" "$environment" "$user"  "$password"
+
+			exit 1
+
+		fi
+
+	else
+
+		echo "bcl2fastq_completed_past for runID - $runID and queueid $queueID and runid fastqStatus '$fastqStatus' --"
+		updateStatus "$queueID" "bcl2fastq_completed_past" "$environment" "$user"  "$password"
+
+fi
+done
+
 
 # run nextseq jobs
 /opt/parallel/bin/parallel --jobs /var/pipelines_"$environment"/run_files/nextseq_jobfile \
-				 --eta \
          -a /var/pipelines_"$environment"/run_files/"nextseq_$currentdate".txt  "/var/pipelines_"$environment"/shell/pipelineThread.sh -e $environment -u $user -p $password -q "
 
 
 sudo rm -f /var/pipelines_"$environment"/run_files/"nextseq_$currentdate".txt
 
-echo "nextseq cron job finished running at - $currentdate"
+echo " $currentdate    INFO  -  NEXTSEQ cron job finished running"
